pyproject.toml
[project]
name = "pytorchlightning-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "efficientnet-pytorch-3d",
    "imagecodecs>=2025.3.30",
    "lightning>=2.5.5",
    "matplotlib>=3.10.7",
    "natsort>=8.4.0",
    "onnx>=1.19.1",
    "onnxruntime-gpu>=1.23.2",
    "onnxscript>=0.5.6",
    "scipy>=1.15.3",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "tifffile>=2025.5.10",
    "toml>=0.10.2",
    "torchsummary>=1.5.1",
    "torchsummaryx>=1.3.0",
    "torchvision>=0.24.0",
]

[tool.uv.sources]
efficientnet-pytorch-3d = { git = "https://github.com/shijianjian/EfficientNet-PyTorch-3D" }
